refactor: Split controllers (events and enrol.) [DHIS2-17972]
