+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/EventQueryAnalyticsControllerTest.java
@@ -30,25 +30,23 @@ package org.hisp.dhis.webapi.controller;
 import static org.hisp.dhis.test.web.WebClientUtils.assertStatus;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.hisp.dhis.feedback.ErrorCode;
 import org.hisp.dhis.jsontree.JsonString;
 import org.hisp.dhis.test.web.HttpStatus;
 import org.hisp.dhis.test.webapi.H2ControllerIntegrationTestBase;
 import org.hisp.dhis.test.webapi.json.domain.JsonGrid;
-import org.hisp.dhis.test.webapi.json.domain.JsonWebMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 /**
- * Tests for the {@link EnrollmentAnalyticsController}.
+ * Tests the {@link EventAggregateAnalyticsController}.
  *
  * <p>The main purpose of this test is not to test the correct business logic but to make sure the
  * controller parameters are recognised correctly.
  *
  * @author Jan Bernitt
  */
-class EnrollmentAnalyticsControllerTest extends H2ControllerIntegrationTestBase {
+class EventQueryAnalyticsControllerTest extends H2ControllerIntegrationTestBase {
 
   private String programId;
 
@@ -78,7 +76,7 @@ class EnrollmentAnalyticsControllerTest extends H2ControllerIntegrationTestBase 
   void testGetQueryJson() {
     JsonGrid grid =
         GET(
-                "/analytics/enrollments/query/{program}?dimension=ou:{unit}&startDate=2019-01-01&endDate=2021-01-01",
+                "/analytics/events/query/{program}?dimension=ou:{unit}&startDate=2019-01-01&endDate=2021-01-01",
                 programId,
                 orgUnitId)
             .content()
@@ -87,26 +85,9 @@ class EnrollmentAnalyticsControllerTest extends H2ControllerIntegrationTestBase 
     assertEquals(
         "My Program", grid.getMetaData().getItems().get(programId).getString("name").string());
     assertEquals(
+        "My Unit", grid.getMetaData().getItems().get(orgUnitId).getString("name").string());
+    assertEquals(
         orgUnitId,
         grid.getMetaData().getDimensions().get("ou").get(0).as(JsonString.class).string());
-  }
-
-  @Test
-  void testBadGrammarException() {
-    /* We know this query will fail since it runs on H2 with postgres specific syntax */
-    JsonWebMessage jsonWebMessage =
-        GET(
-                "/analytics/enrollments/query/{program}?dimension=ou:{unit}&startDate=2019-01-01&endDate=2021-01-01",
-                programId,
-                orgUnitId)
-            .content(HttpStatus.CONFLICT)
-            .as(JsonWebMessage.class);
-
-    assertEquals("Conflict", jsonWebMessage.getHttpStatus());
-    assertEquals(409, jsonWebMessage.getHttpStatusCode());
-    assertEquals("ERROR", jsonWebMessage.getStatus());
-    assertEquals(
-        "Query failed because of a syntax error (SqlState: 90022)", jsonWebMessage.getMessage());
-    assertEquals(ErrorCode.E7145, jsonWebMessage.getErrorCode());
   }
 }
