+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/EnrollmentQueryAnalyticsController.java
@@ -1,0 +1,289 @@
+/*
+ * Copyright (c) 2004-2022, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.webapi.controller;
+
+import static org.hisp.dhis.common.DhisApiVersion.ALL;
+import static org.hisp.dhis.common.DhisApiVersion.DEFAULT;
+import static org.hisp.dhis.common.RequestTypeAware.EndpointAction.QUERY;
+import static org.hisp.dhis.common.cache.CacheStrategy.RESPECT_SYSTEM_SETTING;
+import static org.hisp.dhis.period.PeriodDataProvider.DataSource.DATABASE;
+import static org.hisp.dhis.period.PeriodDataProvider.DataSource.SYSTEM_DEFINED;
+import static org.hisp.dhis.security.Authorities.F_PERFORM_ANALYTICS_EXPLAIN;
+import static org.hisp.dhis.setting.SettingKey.ANALYSIS_RELATIVE_PERIOD;
+import static org.hisp.dhis.setting.SettingKey.ANALYTICS_MAX_LIMIT;
+import static org.hisp.dhis.webapi.utils.ContextUtils.CONTENT_TYPE_CSV;
+import static org.hisp.dhis.webapi.utils.ContextUtils.CONTENT_TYPE_EXCEL;
+import static org.hisp.dhis.webapi.utils.ContextUtils.CONTENT_TYPE_HTML;
+import static org.hisp.dhis.webapi.utils.ContextUtils.CONTENT_TYPE_JSON;
+import static org.hisp.dhis.webapi.utils.ContextUtils.CONTENT_TYPE_XML;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import java.util.List;
+import javax.annotation.Nonnull;
+import javax.servlet.http.HttpServletResponse;
+import lombok.AllArgsConstructor;
+import lombok.SneakyThrows;
+import org.hisp.dhis.analytics.analyze.ExecutionPlanStore;
+import org.hisp.dhis.analytics.dimensions.AnalyticsDimensionsPagingWrapper;
+import org.hisp.dhis.analytics.event.EnrollmentAnalyticsDimensionsService;
+import org.hisp.dhis.analytics.event.EnrollmentAnalyticsService;
+import org.hisp.dhis.analytics.event.EventDataQueryService;
+import org.hisp.dhis.analytics.event.EventQueryParams;
+import org.hisp.dhis.analytics.table.setting.AnalyticsTableSettings;
+import org.hisp.dhis.analytics.util.AnalyticsPeriodCriteriaUtils;
+import org.hisp.dhis.common.DhisApiVersion;
+import org.hisp.dhis.common.DimensionsCriteria;
+import org.hisp.dhis.common.EnrollmentAnalyticsQueryCriteria;
+import org.hisp.dhis.common.EventDataQueryRequest;
+import org.hisp.dhis.common.Grid;
+import org.hisp.dhis.common.OpenApi;
+import org.hisp.dhis.common.RequestTypeAware;
+import org.hisp.dhis.common.RequestTypeAware.EndpointAction;
+import org.hisp.dhis.datavalue.DataValue;
+import org.hisp.dhis.period.PeriodDataProvider;
+import org.hisp.dhis.period.RelativePeriodEnum;
+import org.hisp.dhis.security.RequiresAuthority;
+import org.hisp.dhis.setting.SystemSettingManager;
+import org.hisp.dhis.system.grid.GridUtils;
+import org.hisp.dhis.util.PeriodCriteriaUtils;
+import org.hisp.dhis.webapi.dimension.DimensionFilteringAndPagingService;
+import org.hisp.dhis.webapi.dimension.DimensionMapperService;
+import org.hisp.dhis.webapi.dimension.EnrollmentAnalyticsPrefixStrategy;
+import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
+import org.hisp.dhis.webapi.utils.ContextUtils;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@OpenApi.Document(domain = DataValue.class)
+@Controller
+@ApiVersion({DEFAULT, ALL})
+@RequestMapping("/api/analytics/enrollments/query")
+@AllArgsConstructor
+public class EnrollmentQueryAnalyticsController {
+  @Nonnull private final EventDataQueryService eventDataQueryService;
+
+  @Nonnull private final EnrollmentAnalyticsService analyticsService;
+
+  @Nonnull private final ContextUtils contextUtils;
+
+  @Nonnull private final ExecutionPlanStore executionPlanStore;
+
+  @Nonnull private DimensionFilteringAndPagingService dimensionFilteringAndPagingService;
+
+  @Nonnull private EnrollmentAnalyticsDimensionsService enrollmentAnalyticsDimensionsService;
+
+  @Nonnull private DimensionMapperService dimensionMapperService;
+
+  @Nonnull private final SystemSettingManager systemSettingManager;
+
+  @Nonnull private final PeriodDataProvider periodDataProvider;
+
+  @Nonnull private final AnalyticsTableSettings analyticsTableSettings;
+
+  @RequiresAuthority(anyOf = F_PERFORM_ANALYTICS_EXPLAIN)
+  @GetMapping(
+      value = "/{program}/explain",
+      produces = {APPLICATION_JSON_VALUE, "application/javascript"})
+  public @ResponseBody Grid getExplainQueryJson( // JSON, JSONP
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, true, QUERY);
+
+    Grid grid = analyticsService.getEnrollments(params);
+    contextUtils.configureResponse(response, CONTENT_TYPE_JSON, RESPECT_SYSTEM_SETTING);
+
+    if (params.analyzeOnly()) {
+      String key = params.getExplainOrderId();
+      grid.addPerformanceMetrics(executionPlanStore.getExecutionPlans(key));
+    }
+
+    return grid;
+  }
+
+  @GetMapping(
+      value = "/{program}",
+      produces = {APPLICATION_JSON_VALUE, "application/javascript"})
+  public @ResponseBody Grid getQueryJson( // JSON, JSONP
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(response, CONTENT_TYPE_JSON, RESPECT_SYSTEM_SETTING);
+
+    return analyticsService.getEnrollments(params);
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.xml")
+  public void getQueryXml(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_XML, RESPECT_SYSTEM_SETTING, "enrollments.xml", false);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toXml(grid, response.getOutputStream());
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.xls")
+  public void getQueryXls(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_EXCEL, RESPECT_SYSTEM_SETTING, "enrollments.xls", true);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toXls(grid, response.getOutputStream());
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.xlsx")
+  public void getQueryXlsx(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_EXCEL, RESPECT_SYSTEM_SETTING, "enrollments.xlsx", true);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toXlsx(grid, response.getOutputStream());
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.csv")
+  public void getQueryCsv(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_CSV, RESPECT_SYSTEM_SETTING, "enrollments.csv", true);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toCsv(grid, response.getWriter());
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.html")
+  public void getQueryHtml(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_HTML, RESPECT_SYSTEM_SETTING, "enrollments.html", false);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toHtml(grid, response.getWriter());
+  }
+
+  @SneakyThrows
+  @GetMapping("/{program}.html+css")
+  public void getQueryHtmlCss(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      HttpServletResponse response) {
+    EventQueryParams params = getEventQueryParams(program, criteria, apiVersion, false, QUERY);
+
+    contextUtils.configureResponse(
+        response, CONTENT_TYPE_HTML, RESPECT_SYSTEM_SETTING, "enrollments.html", false);
+    Grid grid = analyticsService.getEnrollments(params);
+    GridUtils.toHtmlCss(grid, response.getWriter());
+  }
+
+  @ResponseBody
+  @GetMapping("/dimensions")
+  public AnalyticsDimensionsPagingWrapper<ObjectNode> getQueryDimensions(
+      @RequestParam String programId,
+      @RequestParam(defaultValue = "*") List<String> fields,
+      DimensionsCriteria dimensionsCriteria,
+      HttpServletResponse response) {
+    contextUtils.configureResponse(response, CONTENT_TYPE_JSON, RESPECT_SYSTEM_SETTING);
+    return dimensionFilteringAndPagingService.pageAndFilter(
+        dimensionMapperService.toDimensionResponse(
+            enrollmentAnalyticsDimensionsService.getQueryDimensionsByProgramId(programId),
+            EnrollmentAnalyticsPrefixStrategy.INSTANCE),
+        dimensionsCriteria,
+        fields);
+  }
+
+  private EventQueryParams getEventQueryParams(
+      @PathVariable String program,
+      EnrollmentAnalyticsQueryCriteria criteria,
+      DhisApiVersion apiVersion,
+      boolean analyzeOnly,
+      EndpointAction endpointAction) {
+    criteria.definePageSize(systemSettingManager.getIntSetting(ANALYTICS_MAX_LIMIT));
+
+    if (endpointAction == QUERY) {
+      AnalyticsPeriodCriteriaUtils.defineDefaultPeriodForCriteria(
+          criteria,
+          periodDataProvider,
+          analyticsTableSettings.getMaxPeriodYearsOffset() == null ? SYSTEM_DEFINED : DATABASE);
+    } else {
+      PeriodCriteriaUtils.defineDefaultPeriodForCriteria(
+          criteria,
+          systemSettingManager.getSystemSetting(
+              ANALYSIS_RELATIVE_PERIOD, RelativePeriodEnum.class));
+    }
+
+    EventDataQueryRequest request =
+        EventDataQueryRequest.builder()
+            .fromCriteria(
+                (EnrollmentAnalyticsQueryCriteria)
+                    criteria
+                        .withEndpointAction(endpointAction)
+                        .withEndpointItem(RequestTypeAware.EndpointItem.ENROLLMENT))
+            .program(program)
+            .apiVersion(apiVersion)
+            .build();
+
+    return eventDataQueryService.getFromRequest(request, analyzeOnly);
+  }
+}
